pyproject.toml
[tool.poetry]
name = "nitter_scraper"
version = "0.3.0"
description = ""
authors = ["dgnsrekt <dgnsrekt@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests-html = "^0.10.0"
pydantic = "^1.6.1"
pendulum = "^2.1.2"
loguru = "^0.5.1"
docker = "^4.3.1"
jinja2 = "^2.11.2"

[tool.poetry.dev-dependencies]
pytest-watch = "^4.2.0"
pytest-regressions = "^2.0.1"
tomlkit = "^0.7.0"
pytest-sugar = "^0.9.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
